+++ src/cli/api.rs
@@ -46,6 +46,7 @@ pub enum Output {
 #[serde(rename_all = "camelCase")]
 pub enum Request {
     FindTitle(request::FindTitle),
+    CheckAppUpdate(request::CheckAppUpdate),
 }
 
 /// A response to an individual request.
@@ -54,6 +55,7 @@ pub enum Request {
 pub enum Response {
     Error(response::Error),
     FindTitle(response::FindTitle),
+    CheckAppUpdate(response::CheckAppUpdate),
 }
 
 pub mod request {
@@ -91,6 +93,11 @@ pub mod request {
         /// With multiple values, they will be checked in the order given.
         pub names: Vec<String>,
     }
+
+    /// Check whether an application update is available.
+    #[derive(Debug, Default, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+    #[serde(default, rename_all = "camelCase")]
+    pub struct CheckAppUpdate {}
 }
 
 pub mod response {
@@ -108,6 +115,22 @@ pub mod response {
     pub struct FindTitle {
         /// Any matching titles found.
         pub titles: BTreeSet<String>,
+    }
+
+    #[derive(Debug, Default, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+    #[serde(default, rename_all = "camelCase")]
+    pub struct CheckAppUpdate {
+        /// An available update.
+        pub update: Option<AppUpdate>,
+    }
+
+    #[derive(Debug, Default, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+    #[serde(default, rename_all = "camelCase")]
+    pub struct AppUpdate {
+        /// New version number.
+        pub version: String,
+        /// Release URL to open in browser.
+        pub url: String,
     }
 }
 
@@ -186,6 +209,19 @@ pub fn process(input: Option<String>, config: &Config, manifest: &Manifest) -> R
 
                 responses.push(Response::FindTitle(response::FindTitle { titles }));
             }
+            Request::CheckAppUpdate(request::CheckAppUpdate {}) => match crate::metadata::Release::fetch_sync() {
+                Ok(release) => {
+                    let update = release.is_update().then(|| response::AppUpdate {
+                        version: release.version.to_string(),
+                        url: release.url,
+                    });
+
+                    responses.push(Response::CheckAppUpdate(response::CheckAppUpdate { update }));
+                }
+                Err(e) => {
+                    responses.push(Response::Error(response::Error { message: e.to_string() }));
+                }
+            },
         }
     }